Pass 3b: Full audit — static classes, directory moves, namespace fixes
- Add `static` to all-static Sharlayan extension classes:
  CutsceneAnimationExtension, DutyFinderBellExtension, GameStateExtension,
  MusicExtension, WeatherExtension
- Extract FFXIVWeatherServiceManual to its own file; make FFXIVWeatherExtensions static
- Add `namespace Chromatics.Localization` to LocalizationManager (was in global namespace)
- Add explicit `internal` to RectangularGradient (was implicit)
- Flatten CustomComparers wrapper class: LayerComparer is now a top-level public class
- Move EnumExtensions.cs from Helpers/ to Extensions/ (it defines C# extension methods);
  update namespace Chromatics.Helpers → Chromatics.Extensions; update all callers
- Move VirtualDeviceModel.cs from Models/ to Forms/vDevices/ (VirtualDevice extends
  UserControl — it is UI code, not a data model); update namespace and all callers

Build: 0 Warning(s), 0 Error(s)
Tests: 78/78 passing
