+++ Chromatics/Extensions/EnumExtensions.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Chromatics.Helpers
+namespace Chromatics.Extensions
 {
     public static class EnumExtensions
     {
