+++ Chromatics/Forms/Fm_MainWindow.cs
@@ -15,7 +15,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using AutoUpdaterDotNET;
-using static Chromatics.Models.VirtualDevice;
+using Chromatics.Forms.vDevices;
+using static Chromatics.Forms.vDevices.VirtualDevice;
 using System.Reflection;
 using System.Windows;
 using Chromatics.Models;
