+++ Chromatics/Forms/Uc_Mappings.cs
@@ -3,6 +3,7 @@ using Chromatics.Enums;
 using Chromatics.Extensions;
 using Chromatics.Helpers;
 using Chromatics.Layers;
+using Chromatics.Localization;
 using Chromatics.Models;
 using Gma.System.MouseKeyHook;
 using HidSharp;
@@ -28,7 +29,8 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Xml.Linq;
-using static Chromatics.Models.VirtualDevice;
+using Chromatics.Forms.vDevices;
+using static Chromatics.Forms.vDevices.VirtualDevice;
 using static System.Windows.Forms.VisualStyles.VisualStyleElement.TextBox;
 using MethodInvoker = System.Windows.Forms.MethodInvoker;
 using Point = System.Drawing.Point;
